requirements.txt
# Core dependencies
mcp>=1.0.0
fastmcp>=0.1.0

# OCR model dependencies
torch>=2.0.0
transformers>=5.0.0
pillow>=10.0.0
pypdfium2>=1.0.0
requests>=2.31.0
vllm>=0.6.0

# Web application dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0

# Optional: For GPU support
# Install separately: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
